pyproject.toml
[project]
name = "pdfchatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.10",
    "langchain-classic>=1.0.2",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.10",
    "openai>=2.26.0",
    "presidio-analyzer>=2.2.361",
    "presidio-anonymizer>=2.2.361",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.2",
    "spacy>=3.7.2",
    "streamlit>=1.55.0",
    "tiktoken>=0.12.0",
]
